Moved all mappings into dedicated file
