+++ lightworks/tomography/utils.py
@@ -15,48 +15,6 @@
 import numpy as np
 from scipy.linalg import sqrtm
 
-from .. import qubit
-from ..sdk.circuit import Circuit
-from ..sdk.state import State
-
-PAULI_MAPPING: dict[str, np.ndarray] = {
-    "I": np.array([[1, 0], [0, 1]]),
-    "X": np.array([[0, 1], [1, 0]]),
-    "Y": np.array([[0, -1j], [1j, 0]]),
-    "Z": np.array([[1, 0], [0, -1]]),
-}
-
-RHO_MAPPING: dict[str, np.ndarray] = {
-    "X+": np.array([[1, 1], [1, 1]]) / 2,
-    "X-": np.array([[1, -1], [-1, 1]]) / 2,
-    "Y+": np.array([[1, -1j], [1j, 1]]) / 2,
-    "Y-": np.array([[1, 1j], [-1j, 1]]) / 2,
-    "Z+": np.array([[1, 0], [0, 0]]),
-    "Z-": np.array([[0, 0], [0, 1]]),
-}
-
-r_transform = qubit.H()
-r_transform.add(qubit.S())
-INPUT_MAPPING: dict[str, tuple[State, Circuit]] = {
-    "X+": (State([1, 0]), qubit.H()),
-    "X-": (State([0, 1]), qubit.H()),
-    "Y+": (State([1, 0]), r_transform),
-    "Y-": (State([0, 1]), r_transform),
-    "Z+": (State([1, 0]), qubit.I()),
-    "Z-": (State([0, 1]), qubit.I()),
-}
-
-_y_measure = Circuit(2)
-_y_measure.add(qubit.S())
-_y_measure.add(qubit.Z())
-_y_measure.add(qubit.H())
-MEASUREMENT_MAPPING = {
-    "X": qubit.H(),
-    "Y": _y_measure,
-    "Z": qubit.I(),
-    "I": qubit.I(),
-}
-
 
 def state_fidelity(rho: np.ndarray, rho_exp: np.ndarray) -> float:
     """
